+++ CHANGES.rst
@@ -12,6 +12,11 @@ Improvements
 
 - Add A6 to page size options (:issue:`3793`)
 
+Bugfixes
+^^^^^^^^
+
+- Fix celery/redis dependency issue (:issue:`3809`)
+
 Version 2.1.7
 -------------